refactor: extract loader props to separate file
