+++ src/Shared/Components/Loader/propsLoader.ts
@@ -1,0 +1,11 @@
+export enum LoaderSize {
+  s = "s",
+  m = "m",
+  l = "l",
+}
+
+export interface LoaderProps {
+  loading?: boolean | undefined;
+  size?: LoaderSize;
+  className?: string;
+}
